optimization
создала два компонента под логику карты и телефона. вынесла объемные svg в компоненты
