+++ src/App.js
@@ -1,7 +1,15 @@
-import React, { useState, useEffect, useRef  } from 'react';
+import React, { useState} from 'react';
+
+import YandexMapContainer from './components/YandexMapContainer'; 
+import PhoneModal from './components/PhoneModal'; 
+
+import Blob from './components/Blob'; 
+import BiegeShape from './components/BiegeShape';
+import StepCircles from './components/StepCircles';
 
 import Header from "./components/Header"
 import Footer from "./components/footer"
+
 import dog1 from './img/dog1.jpg';
 import dog2 from './img/dog2.jpg';
 import dog3 from './img/dog3.jpg';
@@ -16,165 +24,18 @@ import pet6 from './img/pet6.jpg';
 import pet7 from './img/pet7.jpg';
 import pet8 from './img/pet8.jpg';
 
-const PhoneModal = () => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [isMobile, setIsMobile] = useState(false);
-
-  useEffect(() => {
-    const checkScreenSize = () => {
-      setIsMobile(window.innerWidth <= 768);
-    };
-
-    checkScreenSize();
-    window.addEventListener('resize', checkScreenSize);
-
-    return () => window.removeEventListener('resize', checkScreenSize);
-  }, []);
-
-  const openModal = () => setIsModalOpen(true);
-  const closeModal = () => setIsModalOpen(false);
-
-  const handlePhoneClick = () => {
-    const phoneNumber = '+79991234567';
-    
-    if (isMobile) {
-      const userConfirmed = window.confirm('Хотите позвонить по номеру +7 (999) 123-45-67?');
-      if (userConfirmed) window.location.href = `tel:${phoneNumber}`;
-    } else {
-      window.location.href = `tel:${phoneNumber}`;
-    }
-  };
-
-  const handleOverlayClick = (e) => {
-    if (e.target === e.currentTarget) closeModal();
-  };
-
-  useEffect(() => {
-    const handleEscape = (e) => {
-      if (e.key === 'Escape') closeModal();
-    };
-
-    if (isModalOpen) {
-      document.addEventListener('keydown', handleEscape);
-    }
-
-    return () => document.removeEventListener('keydown', handleEscape);
-  }, [isModalOpen]);
-
-  return (
-    <div>
-      <button className="call_now" onClick={openModal}>
-        Позвони уже сейчас
-      </button>
-
-      {isModalOpen && (
-        <div className="modal-overlay" onClick={handleOverlayClick}>
-          <div className="modal-content">
-            <button className="close-button" onClick={closeModal}>&times;</button>
-            <h3>Контактный номер</h3>
-            <div className="phone-number" onClick={handlePhoneClick}>
-              +7 (999) 123-45-67
-            </div>
-            <p className='tup_number'>Нажмите на номер для звонка</p>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-}; // ← Убедитесь, что здесь только одна закрывающая фигурная скобка и точка с запятой
-
 function App() {
   const [showAllPets, setShowAllPets] = useState(false);
-  const mapContainerRef = useRef(null);
-  const mapInstanceRef = useRef(null);
-  const [mapReady, setMapReady] = useState(false);
 
   const toggleShowPets = () => {
     setShowAllPets(!showAllPets);
   };
-  useEffect(() => {
-    // Проверяем, не загружен ли уже API
-    if (window.ymaps) {
-      setMapReady(true);
-      return;
-    }
 
-    // Проверяем, не добавляли ли уже скрипт
-    if (document.querySelector('script[src*="api-maps.yandex.ru"]')) {
-      const checkYmaps = setInterval(() => {
-        if (window.ymaps) {
-          clearInterval(checkYmaps);
-          setMapReady(true);
-        }
-      }, 100);
-      return;
-    }
-
-    // Загружаем API Яндекс.Карт
-    const script = document.createElement('script');
-    script.src = 'https://api-maps.yandex.ru/2.1/?apikey=6ab8ba7c-79c3-4987-bfb4-d2fd4c91b9b3&lang=ru_RU';
-    script.async = true;
-    
-    script.onload = () => {
-      window.ymaps.ready(() => {
-        setMapReady(true);
-      });
-    };
-
-    script.onerror = () => {
-      console.error('Ошибка загрузки Яндекс.Карт');
-    };
-
-    document.body.appendChild(script);
-
-    return () => {
-      // Уничтожаем карту при размонтировании
-      if (mapInstanceRef.current) {
-        mapInstanceRef.current.destroy();
-        mapInstanceRef.current = null;
-      }
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!mapReady || !mapContainerRef.current || mapInstanceRef.current) return;
-
-    // Координаты для ул. Академика Киренского, 26к1
-    const coordinates = [55.994446, 92.797586];
-    
-    try {
-      const map = new window.ymaps.Map(mapContainerRef.current, {
-        center: coordinates,
-        zoom: 16
-      });
-
-      const placemark = new window.ymaps.Placemark(coordinates, {
-        hintContent: 'Наш питомник',
-        balloonContent: 'ул. Академика Киренского, 26к1, Красноярск'
-      }, {
-        preset: 'islands#redDotIcon'
-      });
-
-      map.geoObjects.add(placemark);
-      
-      // Убираем лишние элементы управления
-      map.controls.remove('geolocationControl');
-      map.controls.remove('searchControl');
-      map.controls.remove('trafficControl');
-      map.controls.remove('typeSelector');
-      
-      mapInstanceRef.current = map;
-    } catch (error) {
-      console.error('Ошибка инициализации карты:', error);
-    }
-  }, [mapReady]);
   return (
     <div className="wrapper">
       <Header/>
       <div className="background">
-        <svg className="blob" viewBox="0 0 200 200" xmlns="http://www.w3.org/2000/svg">
-          <path fill="#FCDEC2" d="M40,-74C49.5,-61.1,53.1,-44.1,62.7,-30.2C72.3,-16.3,87.8,-5.4,89.9,7.1C92,19.6,80.6,33.8,66.8,48.6C53,63.4,36.8,78.8,17.6,85.7C-1.6,92.6,-23.8,91,-39.7,78.7C-55.6,66.4,-65.2,43.4,-70.9,20.9C-76.6,-1.7,-78.4,-23.7,-68.3,-38.2C-58.2,-52.7,-36.2,-59.6,-18.5,-69.5C-0.8,-79.4,12.6,-92.2,40,-74Z"/>
-        </svg>
+        <Blob />
       </div>
       <div className="first_block" style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
         <div className="blue_block" ></div>
@@ -182,9 +43,7 @@ function App() {
         <img src={dog1} className="dog1"/>
         <img src={dog2} className="dog2"/>
         <div className="biege_block" style={{textAlign: 'center'}}>
-          <svg className='bieges' style={{marginTop: '1vw'}}  viewBox="0 0 664 282" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M481.381 279.549C410.247 290.781 109.357 241.316 19.4532 226.006C5.48562 223.627 -2.86602 210.181 0.923257 196.437L32.5369 81.7718C38.3231 60.7847 73.7944 48.7971 95.17 52.2179C227.909 73.4603 552.278 -63.5334 569.748 39.446C570.721 45.1773 572.621 50.9696 576.478 55.244C673.418 162.683 741.648 249.233 481.381 279.549Z" fill="#FFEDE1"/>
-          </svg>
+          <BiegeShape />
           <span className="dom">Приют <b>"Тёплый дом"</b></span>
           <span className="dom2">️Подари шанс – подари дом!</span>
         </div>
@@ -244,17 +103,7 @@ function App() {
       <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
         <div className="come_home">
           <img src={dog3} className="dog3"/>
-          <svg className='circles' viewBox="0 0 52 376" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M25.9735 288.551C26.6433 288.551 27.2856 288.817 27.7592 289.29C28.2327 289.764 28.4988 290.406 28.4988 291.076V331.48C28.4988 332.15 28.2327 332.792 27.7592 333.266C27.2856 333.74 26.6433 334.006 25.9735 334.006C25.3038 334.006 24.6615 333.74 24.1879 333.266C23.7143 332.792 23.4482 332.15 23.4482 331.48V291.076C23.4482 290.406 23.7143 289.764 24.1879 289.29C24.6615 288.817 25.3038 288.551 25.9735 288.551Z" fill="black"/>
-            <path d="M8.13867 351.051C8.13867 348.626 8.61631 346.225 9.54432 343.984C10.4723 341.744 11.8325 339.708 13.5473 337.994C15.262 336.279 17.2977 334.919 19.5381 333.991C21.7785 333.063 24.1798 332.585 26.6048 332.585C29.0298 332.585 31.431 333.063 33.6714 333.991C35.9119 334.919 37.9475 336.279 39.6623 337.994C41.377 339.708 42.7372 341.744 43.6652 343.984C44.5932 346.225 45.0709 348.626 45.0709 351.051C45.0709 355.949 43.1253 360.646 39.6623 364.109C36.1992 367.572 31.5023 369.517 26.6048 369.517C21.7073 369.517 17.0103 367.572 13.5473 364.109C10.0842 360.646 8.13867 355.949 8.13867 351.051Z" stroke="black" stroke-width="4.10358" stroke-linecap="round" stroke-linejoin="round"/>
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M25.9735 206.479C26.6433 206.479 27.2856 206.745 27.7592 207.219C28.2327 207.692 28.4988 208.335 28.4988 209.004V249.409C28.4988 250.079 28.2327 250.721 27.7592 251.195C27.2856 251.668 26.6433 251.934 25.9735 251.934C25.3038 251.934 24.6615 251.668 24.1879 251.195C23.7143 250.721 23.4482 250.079 23.4482 249.409V209.004C23.4482 208.335 23.7143 207.692 24.1879 207.219C24.6615 206.745 25.3038 206.479 25.9735 206.479Z" fill="black"/>
-            <path d="M8.13867 268.98C8.13867 266.555 8.61631 264.153 9.54432 261.913C10.4723 259.673 11.8325 257.637 13.5473 255.922C15.262 254.207 17.2977 252.847 19.5381 251.919C21.7785 250.991 24.1798 250.514 26.6048 250.514C29.0298 250.514 31.431 250.991 33.6714 251.919C35.9119 252.847 37.9475 254.207 39.6623 255.922C41.377 257.637 42.7372 259.673 43.6652 261.913C44.5932 264.153 45.0709 266.555 45.0709 268.98C45.0709 273.877 43.1253 278.574 39.6623 282.037C36.1992 285.5 31.5023 287.446 26.6048 287.446C21.7073 287.446 17.0103 285.5 13.5473 282.037C10.0842 278.574 8.13867 273.877 8.13867 268.98Z" stroke="black" stroke-width="4.10358" stroke-linecap="round" stroke-linejoin="round"/>
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M25.9735 44.8611C26.6433 44.8611 27.2856 45.1271 27.7592 45.6007C28.2327 46.0743 28.4988 46.7166 28.4988 47.3864V87.7908C28.4988 88.4606 28.2327 89.1029 27.7592 89.5765C27.2856 90.0501 26.6433 90.3161 25.9735 90.3161C25.3038 90.3161 24.6615 90.0501 24.1879 89.5765C23.7143 89.1029 23.4482 88.4606 23.4482 87.7908V47.3864C23.4482 46.7166 23.7143 46.0743 24.1879 45.6007C24.6615 45.1271 25.3038 44.8611 25.9735 44.8611Z" fill="black"/>
-            <path d="M8.13867 107.362C8.13867 104.937 8.61631 102.536 9.54432 100.295C10.4723 98.0548 11.8325 96.0191 13.5473 94.3043C15.262 92.5896 17.2977 91.2294 19.5381 90.3014C21.7785 89.3734 24.1798 88.8958 26.6048 88.8958C29.0298 88.8958 31.431 89.3734 33.6714 90.3014C35.9119 91.2294 37.9475 92.5896 39.6623 94.3043C41.377 96.0191 42.7372 98.0548 43.6652 100.295C44.5932 102.536 45.0709 104.937 45.0709 107.362C45.0709 112.259 43.1253 116.956 39.6623 120.419C36.1992 123.882 31.5023 125.828 26.6048 125.828C21.7073 125.828 17.0103 123.882 13.5473 120.419C10.0842 116.956 8.13867 112.259 8.13867 107.362Z" stroke="black" stroke-width="4.10358" stroke-linecap="round" stroke-linejoin="round"/>
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M25.9735 124.408C26.6433 124.408 27.2856 124.674 27.7592 125.147C28.2327 125.621 28.4988 126.263 28.4988 126.933V167.337C28.4988 168.007 28.2327 168.649 27.7592 169.123C27.2856 169.597 26.6433 169.863 25.9735 169.863C25.3038 169.863 24.6615 169.597 24.1879 169.123C23.7143 168.649 23.4482 168.007 23.4482 167.337V126.933C23.4482 126.263 23.7143 125.621 24.1879 125.147C24.6615 124.674 25.3038 124.408 25.9735 124.408Z" fill="black"/>
-            <path d="M8.13867 186.908C8.13867 184.483 8.61631 182.082 9.54432 179.841C10.4723 177.601 11.8325 175.565 13.5473 173.851C15.262 172.136 17.2977 170.776 19.5381 169.848C21.7785 168.92 24.1798 168.442 26.6048 168.442C29.0298 168.442 31.431 168.92 33.6714 169.848C35.9119 170.776 37.9475 172.136 39.6623 173.851C41.377 175.565 42.7372 177.601 43.6652 179.841C44.5932 182.082 45.0709 184.483 45.0709 186.908C45.0709 191.806 43.1253 196.502 39.6623 199.966C36.1992 203.429 31.5023 205.374 26.6048 205.374C21.7073 205.374 17.0103 203.429 13.5473 199.966C10.0842 196.502 8.13867 191.806 8.13867 186.908Z" stroke="black" stroke-width="4.10358" stroke-linecap="round" stroke-linejoin="round"/>
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M25.9742 4.87756C14.3516 4.87756 4.93018 14.299 4.93018 25.9216C4.93018 37.5442 14.3516 46.9655 25.9742 46.9655C37.5968 46.9655 47.0182 37.5442 47.0182 25.9216C47.0182 14.299 37.5968 4.87756 25.9742 4.87756ZM25.9742 17.504C23.7417 17.504 21.6006 18.3908 20.022 19.9694C18.4434 21.548 17.5566 23.6891 17.5566 25.9216C17.5566 28.154 18.4434 30.2951 20.022 31.8737C21.6006 33.4523 23.7417 34.3392 25.9742 34.3392C28.2067 34.3392 30.3477 33.4523 31.9263 31.8737C33.5049 30.2951 34.3918 28.154 34.3918 25.9216C34.3918 23.6891 33.5049 21.548 31.9263 19.9694C30.3477 18.3908 28.2067 17.504 25.9742 17.504Z" fill="black"/>
-          </svg>
+          <StepCircles />
           <span className="one"><b>Поедем домой?</b></span>
           <span className="two">Выбери подходящего питомца</span>
           <span className="three">Позвони и договорись о встече</span>
@@ -263,8 +112,6 @@ function App() {
           <span className="six">Привези нового друга домой</span>
           {/* <button className="call_now" onclick="showPhoneModal()">Позвони уже сейчас</button> */}
                <PhoneModal />
-
-
         </div> 
       </div>
 
@@ -277,13 +124,13 @@ function App() {
           <div className="one_love">
           <div className='flower'>
             <svg viewBox="0 0 112 111" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <g clip-path="url(#clip0_3128_157)">
+              <g clipPath="url(#clip0_3128_157)">
                 <mask id="mask0_3128_157" style={{maskType:"luminance", maskUnits:"userSpaceOnUse", x:"0", y:"0", width:"112", height:"111"}}>
                   <path d="M0.772949 0.321899H111.144V110.693H0.772949V0.321899Z" fill="white"/>
-                  <path fill-rule="evenodd" clip-rule="evenodd" d="M66.9454 23.1011L66.644 23.4322C66.2526 23.8612 65.8564 24.2858 65.4554 24.7057L64.8271 23.8779C63.5196 22.1799 62.4159 20.9319 61.2316 19.93C59.6099 18.5631 57.9204 17.7354 55.9592 17.7354C54.8979 17.7354 53.8918 17.9943 52.9452 18.457C50.8736 19.4588 49.3369 21.1356 47.2865 24.0011L46.7771 24.7227C46.3894 24.3435 45.9366 23.8808 45.4187 23.3346L45.1343 23.0374C42.7825 20.5668 42.1585 19.9385 41.1057 19.081C40.1124 18.2702 39.2167 17.7226 38.113 17.4509C36.941 17.1419 35.6956 17.2879 34.6269 17.8596C33.5583 18.4314 32.7457 19.3864 32.3524 20.5328C31.1116 23.9729 30.4809 27.6031 30.4888 31.2601C30.4888 47.9856 39.8789 59.7529 55.9592 59.7529C72.0395 59.7529 81.4295 47.9899 81.4295 31.2601C81.4361 27.6029 80.804 23.9726 79.5617 20.5328C79.1667 19.3852 78.3514 18.43 77.2801 17.8597C76.2087 17.2893 74.961 17.1463 73.7884 17.4594C72.6762 17.7438 71.7933 18.3042 70.8169 19.1362C69.7981 19.9937 69.225 20.605 66.9454 23.1011ZM55.9592 51.2628C45.0409 51.2628 38.979 43.6642 38.979 31.2601C38.979 30.496 39.0143 29.7418 39.0851 28.9975L39.2846 29.2013C41.0166 31.0181 42.0863 32.0539 43.1901 32.9029C45.9409 35.0085 48.7256 35.8278 51.2726 32.9157C51.7311 32.3935 52.1811 31.8035 52.7839 30.9502L53.1023 30.4917L54.1932 28.9423C54.9361 27.898 55.5602 27.1297 56.0144 26.65C56.7717 27.4014 57.4713 28.2088 58.1072 29.0654C57.9162 28.8192 60.442 32.194 61.0787 32.9157C63.6343 35.8405 66.453 34.9958 69.1401 32.8732C70.2014 32.0242 71.2287 30.9927 72.8545 29.2225C72.9111 29.8989 72.9394 30.5781 72.9394 31.2601C72.9394 43.6684 66.8775 51.2628 55.9592 51.2628Z" fill="black"/>
+                  <path fillRule="evenodd" clipRule="evenodd" d="M66.9454 23.1011L66.644 23.4322C66.2526 23.8612 65.8564 24.2858 65.4554 24.7057L64.8271 23.8779C63.5196 22.1799 62.4159 20.9319 61.2316 19.93C59.6099 18.5631 57.9204 17.7354 55.9592 17.7354C54.8979 17.7354 53.8918 17.9943 52.9452 18.457C50.8736 19.4588 49.3369 21.1356 47.2865 24.0011L46.7771 24.7227C46.3894 24.3435 45.9366 23.8808 45.4187 23.3346L45.1343 23.0374C42.7825 20.5668 42.1585 19.9385 41.1057 19.081C40.1124 18.2702 39.2167 17.7226 38.113 17.4509C36.941 17.1419 35.6956 17.2879 34.6269 17.8596C33.5583 18.4314 32.7457 19.3864 32.3524 20.5328C31.1116 23.9729 30.4809 27.6031 30.4888 31.2601C30.4888 47.9856 39.8789 59.7529 55.9592 59.7529C72.0395 59.7529 81.4295 47.9899 81.4295 31.2601C81.4361 27.6029 80.804 23.9726 79.5617 20.5328C79.1667 19.3852 78.3514 18.43 77.2801 17.8597C76.2087 17.2893 74.961 17.1463 73.7884 17.4594C72.6762 17.7438 71.7933 18.3042 70.8169 19.1362C69.7981 19.9937 69.225 20.605 66.9454 23.1011ZM55.9592 51.2628C45.0409 51.2628 38.979 43.6642 38.979 31.2601C38.979 30.496 39.0143 29.7418 39.0851 28.9975L39.2846 29.2013C41.0166 31.0181 42.0863 32.0539 43.1901 32.9029C45.9409 35.0085 48.7256 35.8278 51.2726 32.9157C51.7311 32.3935 52.1811 31.8035 52.7839 30.9502L53.1023 30.4917L54.1932 28.9423C54.9361 27.898 55.5602 27.1297 56.0144 26.65C56.7717 27.4014 57.4713 28.2088 58.1072 29.0654C57.9162 28.8192 60.442 32.194 61.0787 32.9157C63.6343 35.8405 66.453 34.9958 69.1401 32.8732C70.2014 32.0242 71.2287 30.9927 72.8545 29.2225C72.9111 29.8989 72.9394 30.5781 72.9394 31.2601C72.9394 43.6684 66.8775 51.2628 55.9592 51.2628Z" fill="black"/>
                   <path d="M51.7139 59.7529C51.7139 58.627 52.1611 57.5473 52.9572 56.7512C53.7533 55.9551 54.8331 55.5078 55.9589 55.5078C57.0848 55.5078 58.1645 55.9551 58.9606 56.7512C59.7567 57.5473 60.204 58.627 60.204 59.7529V87.3457C60.204 88.4716 59.7567 89.5514 58.9606 90.3475C58.1645 91.1436 57.0848 91.5908 55.9589 91.5908C54.8331 91.5908 53.7533 91.1436 52.9572 90.3475C52.1611 89.5514 51.7139 88.4716 51.7139 87.3457V59.7529Z" fill="black"/>
-                  <path fill-rule="evenodd" clip-rule="evenodd" d="M63.461 68.3065C54.1855 76.092 48.0854 85.9872 53.1115 91.977C58.0612 97.8776 71.7049 94.6429 79.8341 87.8168C87.7639 81.1648 91.3807 71.537 86.4395 65.6491C81.4982 59.7612 71.3907 61.6503 63.461 68.3065ZM59.8654 86.6791C59.812 86.6589 59.7608 86.6333 59.7126 86.6027L59.789 86.7301C59.8201 86.7584 59.8456 86.7414 59.8654 86.6791ZM74.375 81.3134C69.6715 85.2613 61.5804 87.3499 59.8654 86.6791L59.8951 86.5518C60.0564 85.7495 60.5573 84.6118 61.3724 83.2873C63.0874 80.5026 65.9443 77.3061 68.9201 74.81C73.734 70.7687 78.8493 69.8135 79.936 71.1083C81.0228 72.403 79.1931 77.2721 74.375 81.3134Z" fill="black"/>
-                  <path fill-rule="evenodd" clip-rule="evenodd" d="M47.1805 68.3065C56.4602 76.092 62.5561 85.9872 57.53 91.977C52.5802 97.8776 38.9409 94.6429 30.8116 87.8168C22.8776 81.1648 19.265 71.537 24.2063 65.6491C29.1475 59.7612 39.2508 61.6503 47.1805 68.3065ZM50.7803 86.6791L50.9289 86.6027L50.8525 86.7301C50.8214 86.7584 50.7973 86.7414 50.7803 86.6791ZM36.2622 81.3176C40.9658 85.2655 49.0611 87.3541 50.7761 86.6834L50.7464 86.556C50.5851 85.7537 50.0841 84.616 49.2691 83.2916C47.5498 80.5068 44.6972 77.3103 41.7214 74.8142C36.9032 70.7729 31.7879 69.8178 30.7055 71.1125C29.623 72.4073 31.4483 77.2763 36.2665 81.3176" fill="black"/>
+                  <path fillRule="evenodd" clipRule="evenodd" d="M63.461 68.3065C54.1855 76.092 48.0854 85.9872 53.1115 91.977C58.0612 97.8776 71.7049 94.6429 79.8341 87.8168C87.7639 81.1648 91.3807 71.537 86.4395 65.6491C81.4982 59.7612 71.3907 61.6503 63.461 68.3065ZM59.8654 86.6791C59.812 86.6589 59.7608 86.6333 59.7126 86.6027L59.789 86.7301C59.8201 86.7584 59.8456 86.7414 59.8654 86.6791ZM74.375 81.3134C69.6715 85.2613 61.5804 87.3499 59.8654 86.6791L59.8951 86.5518C60.0564 85.7495 60.5573 84.6118 61.3724 83.2873C63.0874 80.5026 65.9443 77.3061 68.9201 74.81C73.734 70.7687 78.8493 69.8135 79.936 71.1083C81.0228 72.403 79.1931 77.2721 74.375 81.3134Z" fill="black"/>
+                  <path fillRule="evenodd" clipRule="evenodd" d="M47.1805 68.3065C56.4602 76.092 62.5561 85.9872 57.53 91.977C52.5802 97.8776 38.9409 94.6429 30.8116 87.8168C22.8776 81.1648 19.265 71.537 24.2063 65.6491C29.1475 59.7612 39.2508 61.6503 47.1805 68.3065ZM50.7803 86.6791L50.9289 86.6027L50.8525 86.7301C50.8214 86.7584 50.7973 86.7414 50.7803 86.6791ZM36.2622 81.3176C40.9658 85.2655 49.0611 87.3541 50.7761 86.6834L50.7464 86.556C50.5851 85.7537 50.0841 84.616 49.2691 83.2916C47.5498 80.5068 44.6972 77.3103 41.7214 74.8142C36.9032 70.7729 31.7879 69.8178 30.7055 71.1125C29.623 72.4073 31.4483 77.2763 36.2665 81.3176" fill="black"/>
                 </mask>
                 <g mask="url(#mask0_3128_157)">
                   <path d="M55.9587 110.693C86.4369 110.693 111.144 85.9859 111.144 55.5076C111.144 25.0294 86.4369 0.321899 55.9587 0.321899C25.4804 0.321899 0.772949 25.0294 0.772949 55.5076C0.772949 85.9859 25.4804 110.693 55.9587 110.693Z" fill="#4D4F50"/>
@@ -318,14 +165,14 @@ function App() {
           <div className="two_love">
             <div className='present_svg'>
               <svg viewBox="0 0 112 111" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <g clip-path="url(#clip0_3128_145)">
+                <g clipPath="url(#clip0_3128_145)">
                   <mask id="mask0_3128_145" style={{maskType:"luminance", maskUnits:"userSpaceOnUse", x:"0", y:"0", width:"112", height:"111"}}>
                     <path d="M0.972168 0.322021H111.344V110.694H0.972168V0.322021Z" fill="white"/>
                     <path d="M51.9126 42.7726H60.4027V89.4682H51.9126V42.7726Z" fill="black"/>
-                    <path fill-rule="evenodd" clip-rule="evenodd" d="M57.6692 42.7641C61.8209 37.9884 58.6116 28.1824 53.0633 23.3642C47.6339 18.6352 40.0267 17.7692 35.909 22.5025C31.7913 27.2357 33.7143 34.6561 39.1438 39.3766C45.4264 44.8357 53.352 47.7308 57.6734 42.7599M51.2761 37.1734C51.4382 36.9835 51.6455 36.8375 51.8789 36.7489C51.6688 36.7319 51.4602 36.6992 51.2549 36.6512C51.3058 36.9272 51.3101 37.1182 51.2634 37.1861L51.2761 37.1734ZM44.7048 32.9708C46.3519 34.4014 48.4701 35.7046 50.2318 36.3456C50.6196 36.4871 50.9592 36.589 51.2507 36.6512C50.996 35.2376 49.5357 31.5487 47.4895 29.77C45.1972 27.7748 42.8072 27.5031 42.3106 28.0762C41.8097 28.6493 42.4125 30.9798 44.7048 32.9708Z" fill="black"/>
-                    <path fill-rule="evenodd" clip-rule="evenodd" d="M55.6739 42.7636C51.5223 37.9879 54.7315 28.1818 60.2756 23.3637C65.7093 18.6389 73.3206 17.7729 77.4341 22.5062C81.5518 27.2394 79.633 34.6598 74.2036 39.3803C67.9209 44.8394 59.9954 47.7345 55.6739 42.7636ZM62.0585 37.1728C61.8976 36.9835 61.6918 36.8376 61.46 36.7483C61.5831 36.7398 61.8081 36.7101 62.084 36.6507C62.033 36.9266 62.0288 37.1177 62.0755 37.1856L62.0585 37.1728ZM68.6299 32.9702C66.987 34.4008 64.8687 35.704 63.1113 36.3451C62.7756 36.4685 62.4325 36.5705 62.084 36.6507C62.3387 35.2371 63.799 31.5481 65.8451 29.7695C68.1374 27.7743 70.5274 27.5026 71.0241 28.0757C71.525 28.6488 70.9222 30.9793 68.6299 32.9702Z" fill="black"/>
+                    <path fillRule="evenodd" clipRule="evenodd" d="M57.6692 42.7641C61.8209 37.9884 58.6116 28.1824 53.0633 23.3642C47.6339 18.6352 40.0267 17.7692 35.909 22.5025C31.7913 27.2357 33.7143 34.6561 39.1438 39.3766C45.4264 44.8357 53.352 47.7308 57.6734 42.7599M51.2761 37.1734C51.4382 36.9835 51.6455 36.8375 51.8789 36.7489C51.6688 36.7319 51.4602 36.6992 51.2549 36.6512C51.3058 36.9272 51.3101 37.1182 51.2634 37.1861L51.2761 37.1734ZM44.7048 32.9708C46.3519 34.4014 48.4701 35.7046 50.2318 36.3456C50.6196 36.4871 50.9592 36.589 51.2507 36.6512C50.996 35.2376 49.5357 31.5487 47.4895 29.77C45.1972 27.7748 42.8072 27.5031 42.3106 28.0762C41.8097 28.6493 42.4125 30.9798 44.7048 32.9708Z" fill="black"/>
+                    <path fillRule="evenodd" clipRule="evenodd" d="M55.6739 42.7636C51.5223 37.9879 54.7315 28.1818 60.2756 23.3637C65.7093 18.6389 73.3206 17.7729 77.4341 22.5062C81.5518 27.2394 79.633 34.6598 74.2036 39.3803C67.9209 44.8394 59.9954 47.7345 55.6739 42.7636ZM62.0585 37.1728C61.8976 36.9835 61.6918 36.8376 61.46 36.7483C61.5831 36.7398 61.8081 36.7101 62.084 36.6507C62.033 36.9266 62.0288 37.1177 62.0755 37.1856L62.0585 37.1728ZM68.6299 32.9702C66.987 34.4008 64.8687 35.704 63.1113 36.3451C62.7756 36.4685 62.4325 36.5705 62.084 36.6507C62.3387 35.2371 63.799 31.5481 65.8451 29.7695C68.1374 27.7743 70.5274 27.5026 71.0241 28.0757C71.525 28.6488 70.9222 30.9793 68.6299 32.9702Z" fill="black"/>
                     <path d="M85.8737 59.7527V68.2428H26.4429V59.7527H85.8737Z" fill="black"/>
-                    <path fill-rule="evenodd" clip-rule="evenodd" d="M26.4423 38.5276H85.8731C86.999 38.5276 88.0787 38.9748 88.8748 39.7709C89.6709 40.567 90.1182 41.6468 90.1182 42.7726V89.4683C90.1182 90.5941 89.6709 91.6739 88.8748 92.47C88.0787 93.2661 86.999 93.7133 85.8731 93.7133H26.4423C25.3165 93.7133 24.2367 93.2661 23.4406 92.47C22.6445 91.6739 22.1973 90.5941 22.1973 89.4683V42.7726C22.1973 41.6468 22.6445 40.567 23.4406 39.7709C24.2367 38.9748 25.3165 38.5276 26.4423 38.5276ZM81.6281 85.2232V47.0177H30.6874V85.2232H81.6281Z" fill="black"/>
+                    <path fillRule="evenodd" clipRule="evenodd" d="M26.4423 38.5276H85.8731C86.999 38.5276 88.0787 38.9748 88.8748 39.7709C89.6709 40.567 90.1182 41.6468 90.1182 42.7726V89.4683C90.1182 90.5941 89.6709 91.6739 88.8748 92.47C88.0787 93.2661 86.999 93.7133 85.8731 93.7133H26.4423C25.3165 93.7133 24.2367 93.2661 23.4406 92.47C22.6445 91.6739 22.1973 90.5941 22.1973 89.4683V42.7726C22.1973 41.6468 22.6445 40.567 23.4406 39.7709C24.2367 38.9748 25.3165 38.5276 26.4423 38.5276ZM81.6281 85.2232V47.0177H30.6874V85.2232H81.6281Z" fill="black"/>
                   </mask>
                   <g mask="url(#mask0_3128_145)">
                     <path d="M56.1579 110.694C86.6362 110.694 111.344 85.986 111.344 55.5078C111.344 25.0295 86.6362 0.322021 56.1579 0.322021C25.6797 0.322021 0.972168 25.0295 0.972168 55.5078C0.972168 85.986 25.6797 110.694 56.1579 110.694Z" fill="#4D4F50"/>
@@ -361,7 +208,7 @@ function App() {
           <div className="three_love">
             <div className='home_svg'>
               <svg viewBox="0 0 112 111" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <g clip-path="url(#clip0_3128_143)">
+                <g clipPath="url(#clip0_3128_143)">
                   <path d="M55.8532 0.322144C25.375 0.322144 0.66748 25.0296 0.66748 55.5079C0.66748 85.9861 25.375 110.694 55.8532 110.694C86.3315 110.694 111.039 85.9861 111.039 55.5079C111.039 25.0296 86.3315 0.322144 55.8532 0.322144ZM55.8532 25.7526L82.7994 46.6965V85.2632H64.2604V61.9817H47.446V85.2632H28.907V46.6965L55.8532 25.7526Z" fill="#4D4F50"/>
                 </g>
                 <defs>
@@ -398,7 +245,7 @@ function App() {
       </div>
 
       <div style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}} id="href_number">
-        <dev className="pink"></dev>
+        <div className="pink"></div>
         <img src={pets} className="pets"/>
         <span className="your"><b>Ваш вклад спасает жизни – не откладывайте добро на потом!</b></span>
 
@@ -424,31 +271,8 @@ function App() {
         <span className="KN"><b>Добрые телефоны</b></span>
       </div>
 
-      <div className="block_map" style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
-        <div className="biege">
-          <svg viewBox="0 0 214 233" fill="none" xmlns="http://www.w3.org/2000/svg" className="leg">
-            <path d="M194.425 47.7522C194.425 47.7522 196.16 55.7421 195.508 57.0121C194.856 58.2821 189.255 65.7676 188.285 65.8301C187.316 65.8926 181.055 67.0336 181.055 67.0336L176.396 58.985L180.71 50.3582L189.448 46.1111L194.425 47.7522ZM173.578 18.7479L174.66 28.1576L169.217 35.9025L162.823 36.3146L159.204 31.737L159.961 21.3514L169.061 13.7302L173.578 18.7479ZM137.797 9.33171L140.466 16.2918L140.668 23.1487L137.489 29.5834L131.472 31.1467L127.148 25.2181L126.524 15.5392C126.524 15.5392 132.503 8.15293 133.744 8.07292C134.986 7.99291 137.794 9.32822 137.794 9.32822M107.138 25.7514L112.27 33.2977C112.27 33.2977 112.076 45.0435 111.204 46.0356C110.332 47.0278 102.295 50.2413 102.295 50.2413L95.4542 43.4303L94.7734 31.0071L99.9671 24.4576L107.138 25.7514ZM155.182 58.162C155.552 59.0741 162.776 69.6829 162.776 69.6829L162.395 83.1217L167.547 95.268L165.861 107.855L154.721 114.728L144.341 111.683L133.276 102.999L118.994 107.768C117.583 108.214 116.082 108.289 114.635 107.984C113.187 107.68 111.843 107.008 110.731 106.032C107.413 103.112 103.981 98.6007 104.134 98.0105C104.384 97.0585 104.513 86.7545 104.513 86.7545C104.513 86.7545 112.114 74.4453 114.95 74.8877C114.95 74.8877 117.817 58.0803 129.404 53.9004C140.988 49.7208 145.975 52.5143 145.975 52.5143L155.182 58.162ZM110.928 134.764C110.928 134.764 118.506 144.002 117.947 144.974C117.388 145.946 114.781 156.549 114.781 156.549C114.781 156.549 106.188 158.813 105.819 157.901C105.449 156.989 101.867 149.68 101.867 149.68L103.334 138.335L110.928 134.764ZM78.3082 116.294L85.5335 126.745L79.6174 140.792L68.6737 136.256L66.3177 126.943L69.1395 118.154L78.3082 116.294ZM43.4092 127.816C43.7787 128.729 52.8441 139.151 52.8441 139.151L50.1859 151.176L42.7474 149.499L36.627 141.844L36.2927 132.416L43.4092 127.816ZM26.865 158.721L34.1387 167.137L36.5368 175.301L33.5325 179.515L24.9475 178.29C24.9475 178.29 18.3204 171.252 18.2603 170.32C18.2002 169.388 18.1977 162.088 18.1977 162.088L26.865 158.721ZM96.0763 170.434C97.0517 174.067 100.628 181.582 100.628 181.582C100.628 181.582 111.047 187.44 111.75 188.645C112.449 189.847 115.772 198.547 115.772 198.547L113.157 208.498L105.395 214.644L87.2225 211.049L75.4811 223.472C75.4811 223.472 65.2679 228.933 59.9575 223.971C54.6472 219.008 52.2913 210.163 52.2913 210.163L56.1497 197.998L54.0371 182.702C54.0371 182.702 59.3245 164.709 69.1516 162.203C78.9714 159.699 94.8439 165.834 96.0763 170.434Z" fill="#FFCC9D"/>
-            <path d="M112.931 74.5779C113.009 73.74 113.123 72.9058 113.271 72.0775C116.079 56.8575 129.865 46.6468 144.073 49.2654C158.281 51.884 167.517 66.3466 164.709 81.5629C164.551 82.3992 164.359 83.219 164.134 84.0222C168.716 88.141 171.155 94.6815 169.892 101.517C168.023 111.663 158.829 118.471 149.36 116.726C142.98 115.55 134.688 104.519 133.907 104.379C133.127 104.238 121.444 111.576 115.067 110.4C105.595 108.652 99.4365 99.0151 101.31 88.8684C102.569 82.0333 107.18 76.7904 112.931 74.5779Z" stroke="black" stroke-width="6.05556" stroke-linecap="round" stroke-linejoin="round"/>
-            <path d="M162.444 39.2055C157.153 37.5417 154.753 30.1868 157.083 22.7778C159.412 15.3689 165.59 10.7115 170.881 12.3752C176.172 14.039 178.573 21.3938 176.243 28.8028C173.913 36.2118 167.735 40.8692 162.444 39.2055Z" stroke="black" stroke-width="6.05556" stroke-linecap="round" stroke-linejoin="round"/>
-            <path d="M179.852 68.8607C175.066 66.0578 174.367 58.354 178.291 51.654C182.215 44.9539 189.276 41.7946 194.062 44.5976C198.848 47.4006 199.547 55.1043 195.623 61.8044C191.699 68.5045 184.638 71.6637 179.852 68.8607Z" stroke="black" stroke-width="6.05556" stroke-linecap="round" stroke-linejoin="round"/>
-            <path d="M93.3469 38.3508C92.2693 30.6614 95.8485 23.8039 101.341 23.0342C106.834 22.2644 112.16 27.874 113.238 35.5634C114.315 43.2528 110.736 50.1103 105.243 50.8801C99.7507 51.6498 94.4244 46.0403 93.3469 38.3508Z" stroke="black" stroke-width="6.05556" stroke-linecap="round" stroke-linejoin="round"/>
-            <path d="M85.6114 126.847C83.9618 119.26 78.2429 114.058 72.8187 115.231C67.3946 116.404 64.3332 123.509 65.9791 131.097C67.6249 138.684 73.3321 143.82 78.7682 142.717C84.2374 141.604 87.2661 134.457 85.6079 126.851M134.823 4.76357C129.659 4.34009 124.097 12.6729 123.663 17.9573C123.027 25.6957 127.053 33.2684 132.582 33.7245C138.111 34.1806 143.107 28.2728 143.742 20.5307C144.378 12.7886 140.348 5.21991 134.823 4.76357ZM54.1741 197.663C53.8329 196.901 53.5193 196.12 53.2331 195.32C48.1394 180.707 55.0684 165.008 68.7079 160.254C82.3475 155.5 97.5307 163.488 102.625 178.105C102.905 178.911 103.133 179.72 103.342 180.526C109.364 181.834 114.718 186.315 117.006 192.876C120.4 202.616 115.781 213.086 106.689 216.255C100.563 218.387 87.9042 212.91 87.1498 213.172C86.3955 213.434 79.8863 225.598 73.7602 227.73C64.6685 230.899 54.5462 225.572 51.1478 215.829C48.8605 209.267 50.2731 202.43 54.1741 197.663Z" stroke="black" stroke-width="6.05556" stroke-linecap="round" stroke-linejoin="round"/>
-            <path d="M109.559 160.513C104.013 160.446 99.5932 154.097 99.6865 146.331C99.7799 138.564 104.352 132.323 109.898 132.39C115.444 132.456 119.864 138.806 119.77 146.572C119.677 154.338 115.105 160.58 109.559 160.513Z" stroke="black" stroke-width="6.05556" stroke-linecap="round" stroke-linejoin="round"/>
-            <path d="M19.2186 175.877C14.4751 169.727 14.1899 161.996 18.5816 158.608C22.9734 155.221 30.379 157.46 35.1225 163.61C39.866 169.759 40.1512 177.491 35.7595 180.878C31.3677 184.266 23.9621 182.027 19.2186 175.877Z" stroke="black" stroke-width="6.05556" stroke-linecap="round" stroke-linejoin="round"/>
-            <path d="M35.0831 144.237C31.8052 137.199 33.2239 129.595 38.2518 127.253C43.2797 124.912 50.013 128.719 53.2909 135.758C56.5688 142.797 55.1501 150.401 50.1222 152.743C45.0943 155.084 38.361 151.276 35.0831 144.237Z" stroke="black" stroke-width="6.05556" stroke-linecap="round" stroke-linejoin="round"/>
-          </svg>
+      <YandexMapContainer />
 
-          <div ref={mapContainerRef} className='map'/>
-
-          <svg width="60vw" height="39vw" viewBox="0 0 1261 946" fill="none" xmlns="http://www.w3.org/2000/svg" className='map_bloop'>
-            <path d="M782.345 940.789C776.114 945.242 767.921 946.781 760.555 944.686L19.3026 733.865C7.54443 730.521 -0.14209 719.249 1.03833 707.082L57.7134 122.845C58.0428 119.45 59.0536 116.081 60.8274 113.166C169.496 -65.3685 726.865 157.955 1018 12.5599C1041.16 0.992554 1067.88 -4.8233 1091.22 6.39075C1268.75 91.7099 1463.3 454.153 782.345 940.789Z" fill="#FCDEC2"/>
-          </svg>
-
-          <span className="seek"><b>ИЩИТЕ НАС ЗДЕСЬ!</b></span>
-
-        </div>
-      </div>
-      <hr></hr>
       <Footer/>
     </div>
   );
