+++ src/components/BiegeShape.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const BiegeShape = () => (
+  <svg className='bieges' style={{marginTop: '1vw'}}  viewBox="0 0 664 282" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path d="M481.381 279.549C410.247 290.781 109.357 241.316 19.4532 226.006C5.48562 223.627 -2.86602 210.181 0.923257 196.437L32.5369 81.7718C38.3231 60.7847 73.7944 48.7971 95.17 52.2179C227.909 73.4603 552.278 -63.5334 569.748 39.446C570.721 45.1773 572.621 50.9696 576.478 55.244C673.418 162.683 741.648 249.233 481.381 279.549Z" fill="#FFEDE1"/>
+  </svg>
+);
+
+export default BiegeShape;
