+++ src/components/Blob.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Blob = () => (
+  <svg className="blob" viewBox="0 0 200 200" xmlns="http://www.w3.org/2000/svg">
+    <path fill="#FCDEC2" d="M40,-74C49.5,-61.1,53.1,-44.1,62.7,-30.2C72.3,-16.3,87.8,-5.4,89.9,7.1C92,19.6,80.6,33.8,66.8,48.6C53,63.4,36.8,78.8,17.6,85.7C-1.6,92.6,-23.8,91,-39.7,78.7C-55.6,66.4,-65.2,43.4,-70.9,20.9C-76.6,-1.7,-78.4,-23.7,-68.3,-38.2C-58.2,-52.7,-36.2,-59.6,-18.5,-69.5C-0.8,-79.4,12.6,-92.2,40,-74Z"/>
+  </svg>
+);
+
+export default Blob;
