+++ src/components/StepCircles.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const StepCircles = () => (
+<svg className='circles' viewBox="0 0 52 376" fill="none" xmlns="http://www.w3.org/2000/svg">
+    <path fillRule="evenodd" clipRule="evenodd" d="M25.9735 288.551C26.6433 288.551 27.2856 288.817 27.7592 289.29C28.2327 289.764 28.4988 290.406 28.4988 291.076V331.48C28.4988 332.15 28.2327 332.792 27.7592 333.266C27.2856 333.74 26.6433 334.006 25.9735 334.006C25.3038 334.006 24.6615 333.74 24.1879 333.266C23.7143 332.792 23.4482 332.15 23.4482 331.48V291.076C23.4482 290.406 23.7143 289.764 24.1879 289.29C24.6615 288.817 25.3038 288.551 25.9735 288.551Z" fill="black"/>
+    <path d="M8.13867 351.051C8.13867 348.626 8.61631 346.225 9.54432 343.984C10.4723 341.744 11.8325 339.708 13.5473 337.994C15.262 336.279 17.2977 334.919 19.5381 333.991C21.7785 333.063 24.1798 332.585 26.6048 332.585C29.0298 332.585 31.431 333.063 33.6714 333.991C35.9119 334.919 37.9475 336.279 39.6623 337.994C41.377 339.708 42.7372 341.744 43.6652 343.984C44.5932 346.225 45.0709 348.626 45.0709 351.051C45.0709 355.949 43.1253 360.646 39.6623 364.109C36.1992 367.572 31.5023 369.517 26.6048 369.517C21.7073 369.517 17.0103 367.572 13.5473 364.109C10.0842 360.646 8.13867 355.949 8.13867 351.051Z" stroke="black" strokeWidth="4.10358" strokeLinecap="round" strokeLinejoin="round"/>
+    <path fillRule="evenodd" clipRule="evenodd" d="M25.9735 206.479C26.6433 206.479 27.2856 206.745 27.7592 207.219C28.2327 207.692 28.4988 208.335 28.4988 209.004V249.409C28.4988 250.079 28.2327 250.721 27.7592 251.195C27.2856 251.668 26.6433 251.934 25.9735 251.934C25.3038 251.934 24.6615 251.668 24.1879 251.195C23.7143 250.721 23.4482 250.079 23.4482 249.409V209.004C23.4482 208.335 23.7143 207.692 24.1879 207.219C24.6615 206.745 25.3038 206.479 25.9735 206.479Z" fill="black"/>
+    <path d="M8.13867 268.98C8.13867 266.555 8.61631 264.153 9.54432 261.913C10.4723 259.673 11.8325 257.637 13.5473 255.922C15.262 254.207 17.2977 252.847 19.5381 251.919C21.7785 250.991 24.1798 250.514 26.6048 250.514C29.0298 250.514 31.431 250.991 33.6714 251.919C35.9119 252.847 37.9475 254.207 39.6623 255.922C41.377 257.637 42.7372 259.673 43.6652 261.913C44.5932 264.153 45.0709 266.555 45.0709 268.98C45.0709 273.877 43.1253 278.574 39.6623 282.037C36.1992 285.5 31.5023 287.446 26.6048 287.446C21.7073 287.446 17.0103 285.5 13.5473 282.037C10.0842 278.574 8.13867 273.877 8.13867 268.98Z" stroke="black" strokeWidth="4.10358" strokeLinecap="round" strokeLinejoin="round"/>
+    <path fillRule="evenodd" clipRule="evenodd" d="M25.9735 44.8611C26.6433 44.8611 27.2856 45.1271 27.7592 45.6007C28.2327 46.0743 28.4988 46.7166 28.4988 47.3864V87.7908C28.4988 88.4606 28.2327 89.1029 27.7592 89.5765C27.2856 90.0501 26.6433 90.3161 25.9735 90.3161C25.3038 90.3161 24.6615 90.0501 24.1879 89.5765C23.7143 89.1029 23.4482 88.4606 23.4482 87.7908V47.3864C23.4482 46.7166 23.7143 46.0743 24.1879 45.6007C24.6615 45.1271 25.3038 44.8611 25.9735 44.8611Z" fill="black"/>
+    <path d="M8.13867 107.362C8.13867 104.937 8.61631 102.536 9.54432 100.295C10.4723 98.0548 11.8325 96.0191 13.5473 94.3043C15.262 92.5896 17.2977 91.2294 19.5381 90.3014C21.7785 89.3734 24.1798 88.8958 26.6048 88.8958C29.0298 88.8958 31.431 89.3734 33.6714 90.3014C35.9119 91.2294 37.9475 92.5896 39.6623 94.3043C41.377 96.0191 42.7372 98.0548 43.6652 100.295C44.5932 102.536 45.0709 104.937 45.0709 107.362C45.0709 112.259 43.1253 116.956 39.6623 120.419C36.1992 123.882 31.5023 125.828 26.6048 125.828C21.7073 125.828 17.0103 123.882 13.5473 120.419C10.0842 116.956 8.13867 112.259 8.13867 107.362Z" stroke="black" strokeWidth="4.10358" strokeLinecap="round" strokeLinejoin="round"/>
+    <path fillRule="evenodd" clipRule="evenodd" d="M25.9735 124.408C26.6433 124.408 27.2856 124.674 27.7592 125.147C28.2327 125.621 28.4988 126.263 28.4988 126.933V167.337C28.4988 168.007 28.2327 168.649 27.7592 169.123C27.2856 169.597 26.6433 169.863 25.9735 169.863C25.3038 169.863 24.6615 169.597 24.1879 169.123C23.7143 168.649 23.4482 168.007 23.4482 167.337V126.933C23.4482 126.263 23.7143 125.621 24.1879 125.147C24.6615 124.674 25.3038 124.408 25.9735 124.408Z" fill="black"/>
+    <path d="M8.13867 186.908C8.13867 184.483 8.61631 182.082 9.54432 179.841C10.4723 177.601 11.8325 175.565 13.5473 173.851C15.262 172.136 17.2977 170.776 19.5381 169.848C21.7785 168.92 24.1798 168.442 26.6048 168.442C29.0298 168.442 31.431 168.92 33.6714 169.848C35.9119 170.776 37.9475 172.136 39.6623 173.851C41.377 175.565 42.7372 177.601 43.6652 179.841C44.5932 182.082 45.0709 184.483 45.0709 186.908C45.0709 191.806 43.1253 196.502 39.6623 199.966C36.1992 203.429 31.5023 205.374 26.6048 205.374C21.7073 205.374 17.0103 203.429 13.5473 199.966C10.0842 196.502 8.13867 191.806 8.13867 186.908Z" stroke="black" strokeWidth="4.10358" strokeLinecap="round" strokeLinejoin="round"/>
+    <path fillRule="evenodd" clipRule="evenodd" d="M25.9742 4.87756C14.3516 4.87756 4.93018 14.299 4.93018 25.9216C4.93018 37.5442 14.3516 46.9655 25.9742 46.9655C37.5968 46.9655 47.0182 37.5442 47.0182 25.9216C47.0182 14.299 37.5968 4.87756 25.9742 4.87756ZM25.9742 17.504C23.7417 17.504 21.6006 18.3908 20.022 19.9694C18.4434 21.548 17.5566 23.6891 17.5566 25.9216C17.5566 28.154 18.4434 30.2951 20.022 31.8737C21.6006 33.4523 23.7417 34.3392 25.9742 34.3392C28.2067 34.3392 30.3477 33.4523 31.9263 31.8737C33.5049 30.2951 34.3918 28.154 34.3918 25.9216C34.3918 23.6891 33.5049 21.548 31.9263 19.9694C30.3477 18.3908 28.2067 17.504 25.9742 17.504Z" fill="black"/>
+</svg>);
+
+export default StepCircles;
