+++ src/components/YandexMapContainer.js
@@ -1,0 +1,124 @@
+import React, { useEffect, useRef, useState } from 'react';
+
+const YandexMapContainer = () => {
+  const mapContainerRef = useRef(null);
+  const mapInstanceRef = useRef(null);
+  const [mapReady, setMapReady] = useState(false);
+
+  // Эффект для загрузки API Яндекс.Карт
+  useEffect(() => {
+    // Функция, которая будет вызвана после готовности API
+    const handleYmapsReady = () => {
+      setMapReady(true);
+    };
+    
+    // 1. Если API уже загружен (ymaps доступен), просто ждем его готовности
+    if (window.ymaps) {
+      window.ymaps.ready(handleYmapsReady);
+      return;
+    }
+
+    // 2. Проверяем, не добавляли ли уже скрипт, чтобы избежать дублирования
+    if (document.querySelector('script[src*="api-maps.yandex.ru"]')) {
+      const checkYmaps = setInterval(() => {
+        if (window.ymaps) {
+          clearInterval(checkYmaps);
+          window.ymaps.ready(handleYmapsReady);
+        }
+      }, 100);
+      return;
+    }
+
+    // 3. Загружаем API Яндекс.Карт
+    const script = document.createElement('script');
+    script.src = 'https://api-maps.yandex.ru/2.1/?apikey=6ab8ba7c-79c3-4987-bfb4-d2fd4c91b9b3&lang=ru_RU';
+    script.async = true;
+    
+    script.onload = () => {
+      // После загрузки скрипта, ждем готовности самого API
+      window.ymaps.ready(handleYmapsReady);
+    };
+
+    script.onerror = () => {
+      console.error('Ошибка загрузки Яндекс.Карт');
+    };
+
+    document.body.appendChild(script);
+  }, []); // Запускаем только при монтировании
+
+  // Эффект для инициализации карты
+  useEffect(() => {
+    // Инициализируем карту только если API готов, контейнер существует и карта еще не создана
+    if (!mapReady || !mapContainerRef.current || mapInstanceRef.current) return;
+
+    // Координаты для ул. Академика Киренского, 26к1
+    const coordinates = [55.994446, 92.797586];
+    
+    try {
+      const map = new window.ymaps.Map(mapContainerRef.current, {
+        center: coordinates,
+        zoom: 16,
+        controls: [] // Убираем все стандартные элементы управления
+      });
+
+      const placemark = new window.ymaps.Placemark(coordinates, {
+        hintContent: 'Наш питомник',
+        balloonContent: 'ул. Академика Киренского, 26к1, Красноярск'
+      }, {
+        preset: 'islands#redDotIcon'
+      });
+
+      map.geoObjects.add(placemark);
+      
+      // Дополнительно удаляем ненужные элементы управления, если нужны специфические
+      map.controls.remove('geolocationControl');
+      map.controls.remove('searchControl');
+      map.controls.remove('trafficControl');
+      map.controls.remove('typeSelector');
+      
+      mapInstanceRef.current = map;
+    } catch (error) {
+      console.error('Ошибка инициализации карты:', error);
+    }
+
+    return () => {
+      // Уничтожаем карту при размонтировании компонента для предотвращения утечек памяти
+      if (mapInstanceRef.current) {
+        mapInstanceRef.current.destroy();
+        mapInstanceRef.current = null;
+      }
+    };
+  }, [mapReady]); // Запускаем при изменении mapReady
+
+  // ВАЖНО: Добавьте стили для div-контейнера карты, чтобы она была видна (например, в вашем CSS)
+  // Я добавил временный style для демонстрации ref
+  return (
+    <div className="block_map" style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+        <div className="biege">
+            <svg viewBox="0 0 214 233" fill="none" xmlns="http://www.w3.org/2000/svg" className="leg">
+                <path d="M194.425 47.7522C194.425 47.7522 196.16 55.7421 195.508 57.0121C194.856 58.2821 189.255 65.7676 188.285 65.8301C187.316 65.8926 181.055 67.0336 181.055 67.0336L176.396 58.985L180.71 50.3582L189.448 46.1111L194.425 47.7522ZM173.578 18.7479L174.66 28.1576L169.217 35.9025L162.823 36.3146L159.204 31.737L159.961 21.3514L169.061 13.7302L173.578 18.7479ZM137.797 9.33171L140.466 16.2918L140.668 23.1487L137.489 29.5834L131.472 31.1467L127.148 25.2181L126.524 15.5392C126.524 15.5392 132.503 8.15293 133.744 8.07292C134.986 7.99291 137.794 9.32822 137.794 9.32822M107.138 25.7514L112.27 33.2977C112.27 33.2977 112.076 45.0435 111.204 46.0356C110.332 47.0278 102.295 50.2413 102.295 50.2413L95.4542 43.4303L94.7734 31.0071L99.9671 24.4576L107.138 25.7514ZM155.182 58.162C155.552 59.0741 162.776 69.6829 162.776 69.6829L162.395 83.1217L167.547 95.268L165.861 107.855L154.721 114.728L144.341 111.683L133.276 102.999L118.994 107.768C117.583 108.214 116.082 108.289 114.635 107.984C113.187 107.68 111.843 107.008 110.731 106.032C107.413 103.112 103.981 98.6007 104.134 98.0105C104.384 97.0585 104.513 86.7545 104.513 86.7545C104.513 86.7545 112.114 74.4453 114.95 74.8877C114.95 74.8877 117.817 58.0803 129.404 53.9004C140.988 49.7208 145.975 52.5143 145.975 52.5143L155.182 58.162ZM110.928 134.764C110.928 134.764 118.506 144.002 117.947 144.974C117.388 145.946 114.781 156.549 114.781 156.549C114.781 156.549 106.188 158.813 105.819 157.901C105.449 156.989 101.867 149.68 101.867 149.68L103.334 138.335L110.928 134.764ZM78.3082 116.294L85.5335 126.745L79.6174 140.792L68.6737 136.256L66.3177 126.943L69.1395 118.154L78.3082 116.294ZM43.4092 127.816C43.7787 128.729 52.8441 139.151 52.8441 139.151L50.1859 151.176L42.7474 149.499L36.627 141.844L36.2927 132.416L43.4092 127.816ZM26.865 158.721L34.1387 167.137L36.5368 175.301L33.5325 179.515L24.9475 178.29C24.9475 178.29 18.3204 171.252 18.2603 170.32C18.2002 169.388 18.1977 162.088 18.1977 162.088L26.865 158.721ZM96.0763 170.434C97.0517 174.067 100.628 181.582 100.628 181.582C100.628 181.582 111.047 187.44 111.75 188.645C112.449 189.847 115.772 198.547 115.772 198.547L113.157 208.498L105.395 214.644L87.2225 211.049L75.4811 223.472C75.4811 223.472 65.2679 228.933 59.9575 223.971C54.6472 219.008 52.2913 210.163 52.2913 210.163L56.1497 197.998L54.0371 182.702C54.0371 182.702 59.3245 164.709 69.1516 162.203C78.9714 159.699 94.8439 165.834 96.0763 170.434Z" fill="#FFCC9D"/>
+                <path d="M112.931 74.5779C113.009 73.74 113.123 72.9058 113.271 72.0775C116.079 56.8575 129.865 46.6468 144.073 49.2654C158.281 51.884 167.517 66.3466 164.709 81.5629C164.551 82.3992 164.359 83.219 164.134 84.0222C168.716 88.141 171.155 94.6815 169.892 101.517C168.023 111.663 158.829 118.471 149.36 116.726C142.98 115.55 134.688 104.519 133.907 104.379C133.127 104.238 121.444 111.576 115.067 110.4C105.595 108.652 99.4365 99.0151 101.31 88.8684C102.569 82.0333 107.18 76.7904 112.931 74.5779Z" stroke="black" strokeWidth="6.05556" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M162.444 39.2055C157.153 37.5417 154.753 30.1868 157.083 22.7778C159.412 15.3689 165.59 10.7115 170.881 12.3752C176.172 14.039 178.573 21.3938 176.243 28.8028C173.913 36.2118 167.735 40.8692 162.444 39.2055Z" stroke="black" strokeWidth="6.05556" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M179.852 68.8607C175.066 66.0578 174.367 58.354 178.291 51.654C182.215 44.9539 189.276 41.7946 194.062 44.5976C198.848 47.4006 199.547 55.1043 195.623 61.8044C191.699 68.5045 184.638 71.6637 179.852 68.8607Z" stroke="black" strokeWidth="6.05556" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M93.3469 38.3508C92.2693 30.6614 95.8485 23.8039 101.341 23.0342C106.834 22.2644 112.16 27.874 113.238 35.5634C114.315 43.2528 110.736 50.1103 105.243 50.8801C99.7507 51.6498 94.4244 46.0403 93.3469 38.3508Z" stroke="black" strokeWidth="6.05556" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M85.6114 126.847C83.9618 119.26 78.2429 114.058 72.8187 115.231C67.3946 116.404 64.3332 123.509 65.9791 131.097C67.6249 138.684 73.3321 143.82 78.7682 142.717C84.2374 141.604 87.2661 134.457 85.6079 126.851M134.823 4.76357C129.659 4.34009 124.097 12.6729 123.663 17.9573C123.027 25.6957 127.053 33.2684 132.582 33.7245C138.111 34.1806 143.107 28.2728 143.742 20.5307C144.378 12.7886 140.348 5.21991 134.823 4.76357ZM54.1741 197.663C53.8329 196.901 53.5193 196.12 53.2331 195.32C48.1394 180.707 55.0684 165.008 68.7079 160.254C82.3475 155.5 97.5307 163.488 102.625 178.105C102.905 178.911 103.133 179.72 103.342 180.526C109.364 181.834 114.718 186.315 117.006 192.876C120.4 202.616 115.781 213.086 106.689 216.255C100.563 218.387 87.9042 212.91 87.1498 213.172C86.3955 213.434 79.8863 225.598 73.7602 227.73C64.6685 230.899 54.5462 225.572 51.1478 215.829C48.8605 209.267 50.2731 202.43 54.1741 197.663Z" stroke="black" strokeWidth="6.05556" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M109.559 160.513C104.013 160.446 99.5932 154.097 99.6865 146.331C99.7799 138.564 104.352 132.323 109.898 132.39C115.444 132.456 119.864 138.806 119.77 146.572C119.677 154.338 115.105 160.58 109.559 160.513Z" stroke="black" strokeWidth="6.05556" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M19.2186 175.877C14.4751 169.727 14.1899 161.996 18.5816 158.608C22.9734 155.221 30.379 157.46 35.1225 163.61C39.866 169.759 40.1512 177.491 35.7595 180.878C31.3677 184.266 23.9621 182.027 19.2186 175.877Z" stroke="black" strokeWidth="6.05556" strokeLinecap="round" strokeLinejoin="round"/>
+                <path d="M35.0831 144.237C31.8052 137.199 33.2239 129.595 38.2518 127.253C43.2797 124.912 50.013 128.719 53.2909 135.758C56.5688 142.797 55.1501 150.401 50.1222 152.743C45.0943 155.084 38.361 151.276 35.0831 144.237Z" stroke="black" strokeWidth="6.05556" strokeLinecap="round" strokeLinejoin="round"/>
+            </svg>
+
+            <div ref={mapContainerRef} className='map'/>        
+                {!mapReady && <div>Загрузка карты...</div>} 
+
+            <svg width="60vw" height="39vw" viewBox="0 0 1261 946" fill="none" xmlns="http://www.w3.org/2000/svg" className='map_bloop'>
+                <path d="M782.345 940.789C776.114 945.242 767.921 946.781 760.555 944.686L19.3026 733.865C7.54443 730.521 -0.14209 719.249 1.03833 707.082L57.7134 122.845C58.0428 119.45 59.0536 116.081 60.8274 113.166C169.496 -65.3685 726.865 157.955 1018 12.5599C1041.16 0.992554 1067.88 -4.8233 1091.22 6.39075C1268.75 91.7099 1463.3 454.153 782.345 940.789Z" fill="#FCDEC2"/>
+            </svg>
+
+            <span className="seek"><b>ИЩИТЕ НАС ЗДЕСЬ!</b></span>
+
+        </div>
+    </div>
+  );
+};
+
+export default YandexMapContainer;
